requirements.txt
# Core ML libraries
transformers>=4.40.0
torch>=2.1.0
accelerate>=0.26.0
safetensors>=0.4.0

# For loading environment variables
python-dotenv>=1.0.1